bugfix: sys/sysctl header deprecated
After configuring all options the build failed with this error:

In file included from src/helper/options.c:38:
/usr/include/x86_64-linux-gnu/sys/sysctl.h:21:2: error: #warning "The <sys/sysctl.h> header is deprecated and will be removed." [-Werror=cpp]
   21 | #warning "The <sys/sysctl.h> header is deprecated and will be removed."
      |  ^~~~~~~
cc1: all warnings being treated as errors

In the code 'src/helper/options.c', the actual call to sysctl is
conditionally:

if defined(CTL_KERN) && defined(KERN_PROC) && defined(KERN_PROC_PATHNAME) /* *BSD */

This conditions however are not placed around the inclusion of the
header file. In 'my' case, working on Linux, I can assume that the
conditions will not be met, hence the sysctl call will never occur. In
that case I should not include the header in the first place.

This actually solved the issue to have it compiled on Ubuntu 19.10.